requirements.txt
httpx>=0.25.2
openai>=1.104.2,<2.0.0
pandas>=2.3.2
pydantic>=2.7.4,<3.0.0
python-dotenv>=1.0.0
tenacity>=8.2.2
pytest>=7.4.2
pytest-asyncio>=0.22.0
